eShop_API/requirements.txt
asgiref==3.7.2
awscli==1.29.66
awsebcli==3.20.10
blessed==1.20.0
boto3==1.28.63
botocore==1.31.66
cement==2.8.2
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.3
Django==4.2.6
django-filter==23.3
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.16
idna==3.4
jmespath==1.0.1
pathspec==0.10.1
Pillow==10.0.1
psycopg2==2.9.9
pyasn1==0.5.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.7.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
stripe==7.0.0
termcolor==1.1.0
typing_extensions==4.8.0
urllib3==1.26.18
wcwidth==0.1.9
